Adds first revision of RateMonitor
This Monitor is designed to monitor the rates of all neurons over
time simulataneously. It therefore creates very large amounts of
data and should be used with care. Note that as for the now the
filter time constant is fixed to be 3x the sampling interval. This
can be changed easily later.