Changes API of ComplexMatrix
This changes the API of ComplexMatrix to make it as similar as possible
to SimpleMatrix and to minimize reprogramming efforts in all files.
Notably the depth of a synaptic state z_value is now always given at the
end of paramter list for functions. Moreover the commit fixes a few
minor bugs.

Also adds a new unsigned type to auryn_definitions.h which is StateID.