Adds functionality to AuditoryBeebGroup
The group simulates a moving Gaussian bump on a constant background
activity. Depending on the settings a fraction of the onset spikes can
be timelocked similar to what is observed in auditory nerve fibres.
The extension now allows to simulate moving bumps of activity, before
one had to select a fixed center for the mean. This is not required
any more.