Added aEIF (adaptive exponential) to auryn.
Ran tests with different parameters and the model appears to work as
expected. The implmentation can probably be optimised to reduce the
number of instructions used in the evolve method.

Another thing I was thinking about was a "set_mode" method that will let
users set the neurons to different modes, like regular spiking, or
bursting and so on? A constructor with this option would also be nice?