Fixes bug in AdEx spike threshold
The spike generation threshold and e_thr need to be different to 'feel'
the exponential part of the model. I am setting the spike generation thr
spk_thr to 0mV now. Beware that together with the forward-euler
integration this could potentially lead to very high voltages when the
model reaches the spike generating singularity.