Implements load/save for Checkers
This adds the Checker state (only RateChecker) so far to the load/save
mechanisms of the network state. Moreover it updates the documentation
of RateChecker to clearly state that the initial rate estimates might be
off due to the moving average. RateCheckers are initialized with a rate
estimate between the maximum and minimum break point to avoid accidental
run breaking due to inaccurate rate estimates. This closes #17