Changes backend of WeightMonitor
This commit removes the references via pointers in the WeightMonitor
backend, by storing the data_index values instead. This should resolve
problems arising with segfaults if a load operation requires to
reallocate the weight matrix (ComplexMatrix). Note however, that now
elements which are being monitored could become invalid during loading.
For instance if a different weight matrix is loaded. The corresponding
data index might still exist (and will hence be recorded from). Best
practice will be to purge all elements from a WeightMonitor and set the
elements to record from after all loading of weight matrices is
finished. Note that this is now only a problem if the structure of the
weight matrix should be changed after the declararion of WeightMonitor.
Fixes #18.