Added a travis file.
This file works on Travis CI - on my account for the time being, but
this can be configured easily.

The e-mail notification is only sent to me at the moment. That can also
be configured.

It uses some ppas since the travis system is
based on an ubuntu 12.04 LTS the packages of which, frankly, are quite
old. For example, the Cmake version is < 2.88 which is the minimum
required for TARGET_INCLUDE_DIRECTORIES.

Next: add update Readme to show travis build status.