Added refractory period functionality to AdEx.
Even though AdEx has the capability to adapt, in cases where the input
current or excitatory conductance is very high, for example in case of
a network of excitatory neurons, the firing rates of the AdEx can be
very very high - which is not biologically plausible - since the
exponential term comes in to play too quickly. To remedy this, a
refractory period must be included.

I'm leaving the default refractory period as 0, which is closest to the
original model, but I've added the functionality required to set a
refractory period if users require it.