Adds new functionality to SpikingGroup
This adds volatile state vectors  to the SpikingGroup class. State
vectors which are declared with a leading underscore in their name (e.g.
"_temp" will not be saved to files and will not be freed upon
termination of the program either. This allows for instance to create a
postsynaptic trace and register it as a volatile state vector using
add_state_vector("_somename", trace). Then this state_vector can be
filtered using for instance post_state_trace, but there won't be a
double free exception in the end because "trace" will already be freed
as a trace and also be saved as such.