Changes strategy for registering Connections
This change to SparseConnection changes the register process. Now
all connections are registered even if they do not have acive
(evolve_locally) targets or destinations. This ensures that connections
which have to do work pre or postsynaptically can do them. Classic
example is for instance STPConnection for which the presynaptic group is
BLOCKLOCKED to a single rank, but no postsynaptic partners are
integrated on this rank. In this case the old behavior would cause
problems. However, this pathologic case would only arise in very few
specific cases.