Moves AurynVector template to its own file
This moves the new code for the vector object to its own file and
keeps the legacy code which now operates on the object in
auryn_definitions. The AurynVector template is forward declared in
auryn_definitions for now. Tests show that with a good compiler we do
not even need to explicitly implement the SIMD instructions manually.
The four loops are apparently detected and vectorized automatically
without notable performance degradation.
The use of the old functions is now discouraged.