Adds experimental support for state variables in SpikingGroup
State variables are SpikingGroup properties that can be set by the user.
Unlike state vectors they are group wide variables and not neuron
specific variables. This commit adds the function
get_state_variable(str) which returns a pointer to a group variable.
Load and save procedures of this haven't been tested yet.