Implements poor man's version of aligned_alloc
Was using aligned_alloc before which is not available on all platforms.
Shifted to boost::alignment::aligned_alloc in the development branch,
but it requires a relatively recent boost library. Will keep v0.8
running on this manual implementation so it compiles with older Boost
libs and keep the neat memory alignemnt for the next release v0.9 ...